refactor(cat-voices): make link wallet dialog private, move learn more button
