+++ catalyst_voices/lib/widgets/buttons/voices_buttons.dart
@@ -1,7 +1,9 @@
 import 'dart:async';
 
 import 'package:catalyst_voices/widgets/buttons/voices_icon_button.dart';
+import 'package:catalyst_voices/widgets/buttons/voices_text_button.dart';
 import 'package:catalyst_voices_assets/catalyst_voices_assets.dart';
+import 'package:catalyst_voices_localization/catalyst_voices_localization.dart';
 import 'package:flutter/material.dart';
 
 class DrawerToggleButton extends StatelessWidget {
@@ -143,6 +145,25 @@ class MoreOptionsButton extends StatelessWidget {
     return VoicesIconButton(
       onTap: onTap,
       child: VoicesAssets.icons.dotsVertical.buildIcon(),
+    );
+  }
+}
+
+/// A "Learn More" button that redirects usually to an external content.
+class VoicesLearnMoreButton extends StatelessWidget {
+  final VoidCallback onTap;
+
+  const VoicesLearnMoreButton({
+    super.key,
+    required this.onTap,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return VoicesTextButton(
+      trailing: VoicesAssets.icons.externalLink.buildIcon(),
+      onTap: onTap,
+      child: Text(context.l10n.learnMore),
     );
   }
 }
